refactor: modularize the listener hierarchy
Introduces a new TestReporter abstraction that is similar to Gradle's
Test listener but allows us to break up each of the test reporting
styles along with extracting the summary and failure information from
the top level Listener.

The PrettyPrintingListener is now nothing more than a filtering orchestrator,
ignoring non-null events and piping them to our reporters.